refactor: refactor extract or predicate
