+++ src/query/sql/src/planner/plans/scalar.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::HashSet;
 use std::hash::Hash;
 
 use common_ast::ast::BinaryOperator;
@@ -36,8 +35,6 @@ pub trait ScalarExpr {
     fn used_columns(&self) -> ColumnSet;
 
     fn is_deterministic(&self) -> bool;
-
-    fn used_tables(&self) -> HashSet<String>;
 
     // TODO: implement this in the future
     // fn outer_columns(&self) -> ColumnSet;
@@ -102,20 +99,6 @@ impl ScalarExpr for Scalar {
             Scalar::FunctionCall(scalar) => scalar.is_deterministic(),
             Scalar::CastExpr(scalar) => scalar.is_deterministic(),
             Scalar::SubqueryExpr(scalar) => scalar.is_deterministic(),
-        }
-    }
-
-    fn used_tables(&self) -> HashSet<String> {
-        match self {
-            Scalar::BoundColumnRef(scalar) => scalar.used_tables(),
-            Scalar::ConstantExpr(scalar) => scalar.used_tables(),
-            Scalar::AndExpr(scalar) => scalar.used_tables(),
-            Scalar::OrExpr(scalar) => scalar.used_tables(),
-            Scalar::ComparisonExpr(scalar) => scalar.used_tables(),
-            Scalar::AggregateFunction(scalar) => scalar.used_tables(),
-            Scalar::FunctionCall(scalar) => scalar.used_tables(),
-            Scalar::CastExpr(scalar) => scalar.used_tables(),
-            Scalar::SubqueryExpr(scalar) => scalar.used_tables(),
         }
     }
 }
@@ -302,15 +285,6 @@ impl ScalarExpr for BoundColumnRef {
     fn is_deterministic(&self) -> bool {
         true
     }
-
-    fn used_tables(&self) -> HashSet<String> {
-        let mut table_set = HashSet::new();
-        if let Some(table_name) = &self.column.table_name {
-            table_set.insert(table_name.to_string());
-            return table_set;
-        }
-        table_set
-    }
 }
 
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
@@ -331,10 +305,6 @@ impl ScalarExpr for ConstantExpr {
 
     fn is_deterministic(&self) -> bool {
         true
-    }
-
-    fn used_tables(&self) -> HashSet<String> {
-        HashSet::new()
     }
 }
 
@@ -359,13 +329,6 @@ impl ScalarExpr for AndExpr {
     fn is_deterministic(&self) -> bool {
         self.left.is_deterministic() && self.right.is_deterministic()
     }
-
-    fn used_tables(&self) -> HashSet<String> {
-        let mut table_set = HashSet::new();
-        table_set.extend(self.left.used_tables());
-        table_set.extend(self.right.used_tables());
-        table_set
-    }
 }
 
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
@@ -388,13 +351,6 @@ impl ScalarExpr for OrExpr {
 
     fn is_deterministic(&self) -> bool {
         self.left.is_deterministic() && self.right.is_deterministic()
-    }
-
-    fn used_tables(&self) -> HashSet<String> {
-        let mut table_set = HashSet::new();
-        table_set.extend(self.left.used_tables());
-        table_set.extend(self.right.used_tables());
-        table_set
     }
 }
 
@@ -476,13 +432,6 @@ impl ScalarExpr for ComparisonExpr {
             && self.left.is_deterministic()
             && self.right.is_deterministic()
     }
-
-    fn used_tables(&self) -> HashSet<String> {
-        let mut table_set = HashSet::new();
-        table_set.extend(self.left.used_tables());
-        table_set.extend(self.right.used_tables());
-        table_set
-    }
 }
 
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
@@ -511,14 +460,6 @@ impl ScalarExpr for AggregateFunction {
 
     fn is_deterministic(&self) -> bool {
         false
-    }
-
-    fn used_tables(&self) -> HashSet<String> {
-        let mut table_set = HashSet::new();
-        for scalar in self.args.iter() {
-            table_set.extend(scalar.used_tables());
-        }
-        table_set
     }
 }
 
@@ -550,14 +491,6 @@ impl ScalarExpr for FunctionCall {
             .map_or(false, |feature| feature.is_deterministic)
             && self.arguments.iter().all(|arg| arg.is_deterministic())
     }
-
-    fn used_tables(&self) -> HashSet<String> {
-        let mut table_set = HashSet::new();
-        for scalar in self.arguments.iter() {
-            table_set.extend(scalar.used_tables());
-        }
-        table_set
-    }
 }
 
 #[derive(Clone, PartialEq, Eq, Hash, Debug)]
@@ -578,12 +511,6 @@ impl ScalarExpr for CastExpr {
 
     fn is_deterministic(&self) -> bool {
         self.argument.is_deterministic()
-    }
-
-    fn used_tables(&self) -> HashSet<String> {
-        let mut table_set = HashSet::new();
-        table_set.extend(self.argument.used_tables());
-        table_set
     }
 }
 
@@ -644,10 +571,5 @@ impl ScalarExpr for SubqueryExpr {
 
     fn is_deterministic(&self) -> bool {
         false
-    }
-
-    fn used_tables(&self) -> HashSet<String> {
-        // Don't care used tables in subquery
-        HashSet::new()
     }
 }
